(edit) add a new helper formatFacetNames (previously in the useSearchManager hook) and add the transformationHumanReadable in config/options
